main/qemu: Add APKBUILD
This is the source APK for QEMU. It builds:
- qemu-system-aarch64
full arm64 emulator
- qemu-system-x86_64
full x86_64 emulator
- qemu-aarch64
Statically linked qemu user to chroot into an arm64 rootfs on an x86
machine. Used for the arm64 builds on x86 hardware.

It is also building and packaging qemu-img, qemu-io and qemu-nbd.
qemu-img is useful, the others are unnecessary but they are built
automatically.

Signed-off-by: Rajiv Ranganath <rajiv.ranganath@atihita.com>
Acked-by: Stefano Stabellini <sstabellini@kernel.org>
Signed-off-by: Stefano Stabellini <sstabellini@kernel.org>