scripts: Disable cross-building base system
In ViryaOS we plan to only cross-build the Linux kernel and the Xen
hypervisor. We don't plan to cross-build any userspace components. Thus,
we don't need many of the cross-building tools that bootstrap.sh
installs by default. Moreover, some of them are not even available in
upstream Alpine Linux repositories. Just remove them.

Signed-off-by: Rajiv Ranganath <rajiv.ranganath@atihita.com>
Acked-by: Stefano Stabellini <sstabellini@kernel.org>
Signed-off-by: Stefano Stabellini <sstabellini@kernel.org>